requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.22
alembic>=1.12.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
psycopg2-binary>=2.9.9
email-validator>=2.0.0
celery>=5.3.4
redis>=5.0.1
websockets>=12.0
bcrypt>=4.0.1
python-dotenv>=1.0.0
asyncpg>=0.28.0
